refactor: auth user type
