+++ src/auth/strategy/local.strategy.ts
@@ -5,7 +5,7 @@ import { validateOrReject } from 'class-validator';
 import { Strategy } from 'passport-local';
 import ExceptionMessages from '#exception/exception-message.js';
 import { AuthService } from '#auth/auth.service.js';
-import { LoginDTO } from '#auth/type/auth.dto.js';
+import { LoginDTO } from '#auth/dto/auth.dto.js';
 
 @Injectable()
 export class LocalStrategy extends PassportStrategy(Strategy) {
