+++ src/user/type/user.type.ts
@@ -1,11 +1,16 @@
 import { PartialType } from '@nestjs/mapped-types';
 import { Role } from '@prisma/client';
 
-export class CreateUser {
+export interface CreateUser {
   nickname: string;
   email: string;
   password: string;
   role: Role;
 }
 
-export class PatchUser extends PartialType(CreateUser) {}
+export interface PatchUser {
+  nickname?: string;
+  email?: string;
+  password?: string;
+  role?: Role;
+}
